requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
astroid=2.4.1=py36_0
blas=1.0=mkl
ca-certificates=2020.1.1=0
certifi=2020.4.5.1=py36_0
cffi=1.14.0=py36he30daa8_1
click=7.1.2=pypi_0
cuda90=1.0=h6433d27_0
freetype=2.9.1=h8a8886c_1
h5py=2.10.0=py36h7918eee_0
hdf5=1.10.4=hb1b8bf9_0
intel-openmp=2020.1=217
isort=4.3.21=py36_0
jpeg=9b=h024ee3a_2
lazy-object-proxy=1.4.3=py36h7b6447c_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.3=he6710b0_1
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
mccabe=0.6.1=py36_1
mkl=2020.1=217
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.0.15=py36ha843d7b_0
mkl_random=1.1.0=py36hd6b4f25_0
ncurses=6.2=he6710b0_1
nltk=3.4.5=py36_0
numpy=1.18.1=py36h4f9e942_0
numpy-base=1.18.1=py36hde5b4d6_1
olefile=0.46=py36_0
openssl=1.1.1g=h7b6447c_0
pillow=6.1.0=pypi_0
pip=20.0.2=py36_3
pycparser=2.20=py_0
pylint=2.5.2=py36_0
python=3.6.10=h7579374_2
pytorch=0.3.0=py36_cuda9.0.176_cudnn7.0.3hdc18817_4
pyyaml=5.3.1=pypi_0
readline=8.0=h7b6447c_0
scipy=1.1.0=pypi_0
setuptools=46.4.0=py36_0
six=1.14.0=py36_0
skipthoughts=0.0.1=pypi_0
sqlite=3.31.1=h62c20be_1
tk=8.6.8=hbc83047_0
toml=0.10.0=py36h28b3542_0
torchsummary=1.5.1=pypi_0
torchvision=0.2.0=py36h17b6947_1
typed-ast=1.4.1=py36h7b6447c_0
wheel=0.34.2=py36_0
wrapt=1.11.2=py36h7b6447c_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
